backend/requirements.txt
django
djangorestframework
psycopg2-binary
python-decouple
django-cors-headers
gunicorn
python-dotenv
pandas
pandapower 
